🎀 AST lowering typechecks verbs
finally, tying the ribbon together. this is a large rewrite of our
parsing / AST lowering code in `j::r::parse::{parse,parse_}`.

importantly, now that `[` and `]` mean dyadic/monadic verbs are
ambiguous based only on the character, we now peek at the left-hand side
of an expression. if there is a noun on the other side (e.g. an array)
then we interpret it as a dyad.

other small changes:

* `A: TryFrom<Vec<I>>`
* `N: From<SY>`
* `N: TryFrom<Vec<I>>`